libcrypto: Fix integer comparison in evp_cipher_aead_encrypt()
src/libcrypto.c:773:27: warning: comparison of integer expressions of
different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’}
[-Wsign-compare] <--[cc]
     if (rc != 1 || outlen != len - aadlen) {
                           ^~

+++ src/libcrypto.c
@@ -770,7 +770,7 @@ evp_cipher_aead_encrypt(struct ssh_cipher_struct *cipher,
                            &outlen,
                            (unsigned char *)in + aadlen,
                            (int)len - aadlen);
-    if (rc != 1 || outlen != len - aadlen) {
+    if (rc != 1 || outlen != (int)len - aadlen) {
         SSH_LOG(SSH_LOG_WARNING, "EVP_EncryptUpdate failed");
         return;
     }
